requirements.txt
flask==3.0.0
tensorflow==2.16.1
pillow==10.1.0
numpy==1.24.3
werkzeug==3.0.1
